Refactor/WQ-172 구조 리팩토링 (#42)
* refactor: 소셜로그인 dto 불필요한 필드 삭제

* refactor: domain 폴더 통합

* style: 불필요한 주석 제거
