+++ src/main/kotlin/com/wq/auth/api/controller/auth/request/GoogleSocialLoginRequestDto.kt
@@ -1,6 +1,6 @@
 package com.wq.auth.api.controller.auth.request
 
-import com.wq.auth.domain.auth.request.GoogleSocialLoginRequest
+import com.wq.auth.api.domain.auth.request.GoogleSocialLoginRequest
 import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.NotBlank
