+++ src/main/kotlin/com/wq/auth/api/domain/auth/SocialLoginService.kt
@@ -1,21 +1,19 @@
-package com.wq.auth.domain.auth
+package com.wq.auth.api.domain.auth
 
-import com.wq.auth.api.domain.auth.AuthProviderRepository
-import com.wq.auth.api.domain.auth.RefreshTokenRepository
 import com.wq.auth.api.domain.member.MemberRepository
 import com.wq.auth.api.domain.auth.entity.AuthProviderEntity
 import com.wq.auth.api.domain.member.entity.MemberEntity
 import com.wq.auth.api.domain.auth.entity.ProviderType
 import com.wq.auth.api.domain.auth.entity.RefreshTokenEntity
-import com.wq.auth.domain.auth.request.SocialLoginRequest
-import com.wq.auth.domain.auth.response.SocialLoginResult
+import com.wq.auth.api.domain.auth.request.SocialLoginRequest
+import com.wq.auth.api.domain.auth.response.SocialLoginResult
 import com.wq.auth.api.external.oauth.GoogleOAuthClient
 import com.wq.auth.api.external.oauth.KakaoOAuthClient
 import com.wq.auth.api.external.oauth.NaverOAuthClient
-import com.wq.auth.domain.auth.request.OAuthAuthCodeRequest
-import com.wq.auth.domain.oauth.OAuthUser
-import com.wq.auth.domain.oauth.error.SocialLoginException
-import com.wq.auth.domain.oauth.error.SocialLoginExceptionCode
+import com.wq.auth.api.domain.auth.request.OAuthAuthCodeRequest
+import com.wq.auth.api.domain.oauth.OAuthUser
+import com.wq.auth.api.domain.oauth.error.SocialLoginException
+import com.wq.auth.api.domain.oauth.error.SocialLoginExceptionCode
 import com.wq.auth.security.jwt.JwtProvider
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.stereotype.Service
@@ -40,7 +38,7 @@ class SocialLoginService(
     private val memberRepository: MemberRepository,
     private val authProviderRepository: AuthProviderRepository,
     private val jwtProvider: JwtProvider,
-    private val refreshTokenRepository: RefreshTokenRepository
+    private val refreshTokenRepository: RefreshTokenRepository,
 ) {
     private val log = KotlinLogging.logger {}
 
@@ -74,7 +72,6 @@ class SocialLoginService(
             OAuthAuthCodeRequest(
                 request.authCode,
                 request.codeVerifier,
-                request.redirectUri
             )
         )
 
@@ -118,8 +115,7 @@ class SocialLoginService(
         val oauthUser = googleOAuthClient.getUserFromAuthCode(
             OAuthAuthCodeRequest(
                 authCode = request.authCode,
-                codeVerifier = request.codeVerifier,
-                redirectUri = request.redirectUri
+                codeVerifier = request.codeVerifier
             )
         )
 
@@ -164,8 +160,7 @@ class SocialLoginService(
             OAuthAuthCodeRequest(
                 authCode = request.authCode,
                 state = request.state!!,      // 네이버는 state 사용
-                codeVerifier = request.codeVerifier,
-                redirectUri = request.redirectUri
+                codeVerifier = request.codeVerifier
             )
         )
