+++ src/main/kotlin/com/wq/auth/api/domain/auth/request/NaverSocialLoginRequest.kt
@@ -1,8 +1,7 @@
-package com.wq.auth.domain.auth.request
+package com.wq.auth.api.domain.auth.request
 
 data class NaverSocialLoginRequest(
         val authCode: String,
         val state: String,
         val codeVerifier: String,
-        val redirectUri: String? = null
 )
