+++ src/main/kotlin/com/wq/auth/api/domain/auth/request/OAuthAuthCodeRequest.kt
@@ -1,8 +1,7 @@
-package com.wq.auth.domain.auth.request
+package com.wq.auth.api.domain.auth.request
 
 data class OAuthAuthCodeRequest(
     val authCode: String,
     val codeVerifier: String,
     val state: String? = null,  // 네이버만 사용
-    val redirectUri: String? = null
 )
