+++ src/main/kotlin/com/wq/auth/api/domain/auth/request/SocialLoginRequest.kt
@@ -1,4 +1,4 @@
-package com.wq.auth.domain.auth.request
+package com.wq.auth.api.domain.auth.request
 
 import com.wq.auth.api.domain.auth.entity.ProviderType
 
@@ -7,20 +7,16 @@ import com.wq.auth.api.domain.auth.entity.ProviderType
  *
  * 각 소셜 제공자별 OAuth2 방식을 지원합니다:
  * - Google/Kakao: PKCE 방식 (codeVerifier 사용)
- * - Naver: 전통적인 OAuth2 방식 (state 사용)
+ * - Naver: 전통적인 OAuth2 방식 (state 사용) + PKCE
  *
  * @param authCode 소셜 제공자로부터 받은 인가 코드
- * @param codeVerifier PKCE 검증용 코드 검증자 (Google, Kakao용 - 선택사항)
+ * @param codeVerifier PKCE 검증용 코드 검증자
  * @param state CSRF 방지용 상태 값 (Naver용 - 선택사항)
- * @param grantType OAuth2 그랜트 타입 (기본값: "authorization_code")
  * @param providerType 소셜 로그인 제공자 타입
- * @param redirectUri 리다이렉트 URI (선택사항)
  */
 data class SocialLoginRequest(
     val authCode: String,
     val codeVerifier: String,
     val state: String? = null,
-    val grantType: String = "authorization_code",
     val providerType: ProviderType,
-    val redirectUri: String? = null
 )
