+++ src/main/kotlin/com/wq/auth/api/external/oauth/NaverOAuthClient.kt
@@ -3,11 +3,11 @@ package com.wq.auth.api.external.oauth
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.wq.auth.api.domain.auth.entity.ProviderType
 import com.wq.auth.api.external.oauth.dto.NaverUserInfoResponse
-import com.wq.auth.domain.auth.request.OAuthAuthCodeRequest
-import com.wq.auth.domain.oauth.OAuthClient
-import com.wq.auth.domain.oauth.OAuthUser
-import com.wq.auth.domain.oauth.error.SocialLoginException
-import com.wq.auth.domain.oauth.error.SocialLoginExceptionCode
+import com.wq.auth.api.domain.auth.request.OAuthAuthCodeRequest
+import com.wq.auth.api.domain.oauth.OAuthClient
+import com.wq.auth.api.domain.oauth.OAuthUser
+import com.wq.auth.api.domain.oauth.error.SocialLoginException
+import com.wq.auth.api.domain.oauth.error.SocialLoginExceptionCode
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.http.*
 import org.springframework.stereotype.Component
@@ -36,8 +36,6 @@ class NaverOAuthClient(
      *
      * @param authorizationCode Naver로부터 받은 인가 코드
      * @param state CSRF 방지용 상태 값
-     * @param redirectUri 리다이렉트 URI
-<
      * @param codeVerifier PKCE 검증용 코드 검증자
      * @return Naver 액세스 토큰
      * @throws SocialLoginException 토큰 획득 실패 시
@@ -46,10 +44,9 @@ class NaverOAuthClient(
         authorizationCode: String,
         state: String,
         codeVerifier: String,
-        redirectUri: String? = null
     ): String {
         log.info { "Naver 액세스 토큰 요청 시작" }
-        log.info { "redirectUri: ${redirectUri ?: naverOAuthProperties.redirectUri}" }
+        log.info { "redirectUri: ${ naverOAuthProperties.redirectUri}" }
 
         val headers = HttpHeaders().apply {
             contentType = MediaType.APPLICATION_FORM_URLENCODED
@@ -62,7 +59,7 @@ class NaverOAuthClient(
             add("code", authorizationCode)
             add("grant_type", "authorization_code")
             add("state", state)
-            add("redirect_uri", redirectUri ?: naverOAuthProperties.redirectUri)
+            add("redirect_uri", naverOAuthProperties.redirectUri)
         }
 
         val request = HttpEntity(body, headers)
@@ -157,13 +154,13 @@ class NaverOAuthClient(
      *
      * @param authCode Naver로부터 받은 인가 코드
      * @param state CSRF 방지용 상태 값
-     * @param redirectUri 리다이렉트 URI
+     * @param codeVerifier PKCE 검증용 코드 검증자
      * @return 도메인 사용자 정보
      */
     override fun getUserFromAuthCode(req: OAuthAuthCodeRequest): OAuthUser {
         log.info { "Naver AuthCode 요청 시작" }
-        log.info { "redirectUri: ${req.redirectUri ?: naverOAuthProperties.redirectUri}" }
-        val accessToken = getAccessToken(req.authCode, req.state!!, req.codeVerifier, req.redirectUri)
+        log.info { "redirectUri: ${naverOAuthProperties.redirectUri}" }
+        val accessToken = getAccessToken(req.authCode, req.state!!, req.codeVerifier)
         val naverUserInfo = getUserInfo(accessToken)
 
         return OAuthUser(
@@ -182,16 +179,14 @@ class NaverOAuthClient(
      * @param authorizationCode Naver로부터 받은 인가 코드
      * @param state CSRF 방지용 상태 값
      * @param codeVerifier PKCE 검증용 코드 검증자
-     * @param redirectUri 리다이렉트 URI
      * @return Naver 사용자 정보
      */
     fun getUserInfoFromAuthCode(
         authorizationCode: String,
         state: String,
         codeVerifier: String,
-        redirectUri: String? = null
     ): NaverUserInfoResponse {
-        val accessToken = getAccessToken(authorizationCode, state, codeVerifier, redirectUri)
+        val accessToken = getAccessToken(authorizationCode, state, codeVerifier)
         return getUserInfo(accessToken)
     }
 }
